traffic_lights_web/requirements.txt
redis==5.2.0
pyserial==3.5
django==5.1.2
